app/app_service/requirements.txt
fastapi
loguru
uvicorn
requests
python-dotenv